+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Initial objects expressions support.
+- Allow to use Add operator with strings. Implicit conversion from integer, float and numeric types when concat with string.
 
 ### Changed